+++ src/com/intellij/idea/plugin/hybris/polyglotQuery/injection/PolyglotQueryInjector.kt
@@ -18,6 +18,7 @@
 package com.intellij.idea.plugin.hybris.polyglotQuery.injection
 
 import com.intellij.idea.plugin.hybris.polyglotQuery.injection.impl.PolyglotQueryToImpexInjectorProvider
+import com.intellij.idea.plugin.hybris.polyglotQuery.injection.impl.PolyglotQueryToJavaInjectorProvider
 import com.intellij.psi.InjectedLanguagePlaces
 import com.intellij.psi.LanguageInjector
 import com.intellij.psi.PsiLanguageInjectionHost
@@ -30,5 +31,7 @@ class PolyglotQueryInjector : LanguageInjector {
     ) {
         PolyglotQueryToImpexInjectorProvider.instance
             ?.inject(host, injectionPlacesRegistrar)
+            ?: PolyglotQueryToJavaInjectorProvider.instance
+                ?.inject(host, injectionPlacesRegistrar)
     }
 }